# dns/dnsmessage: don't include bytes after name.Length in the compression map