--- a/dns/dnsmessage/message.go
+++ b/dns/dnsmessage/message.go
@@ -2003,6 +2003,8 @@ func (n *Name) pack(msg []byte, compression map[string]int, compressionOff int) 
 			// offset can be stored in the available 14 bytes.
 			if len(msg) <= int(^uint16(0)>>2) {
 				if nameAsStr == "" {
+					// allocate n.Data on the heap once, to avoid allocating it
+					// multiple times (for next labels).
 					nameAsStr = string(n.Data[:n.Length])
 				}
 				compression[nameAsStr[i:]] = len(msg) - compressionOff